// frontend/src/app/api/models/entrada-indice.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
export interface EntradaIndice {

  /**
   * Región (destinos) o categoría (guías) para agrupar en SCR-022.
   */
  agrupacion?: (RefTaxonomia | null);
  publicado_actualizado_en: string;
  slug: Slug;
  tipo: 'DESTINO' | 'ITINERARIO' | 'TIPO' | 'GUIA' | 'CATEGORIA_GUIA' | 'COLECCION' | 'PAGINA';
  titulo: string;
}