// frontend/src/app/api/models/meses.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DestinoTarjeta } from '../models/destino-tarjeta';
import { Mes } from '../models/mes';
export interface Meses {
  meses: Array<{
'mes': Mes;
'nombre': string;
'slug': 'enero' | 'febrero' | 'marzo' | 'abril' | 'mayo' | 'junio' | 'julio' | 'agosto' | 'septiembre' | 'octubre' | 'noviembre' | 'diciembre';
'numero_destinos': number;
'destacado': (DestinoTarjeta | null);
}>;
}